+++ test/test_train_cifar.py
@@ -24,9 +24,9 @@ import torch.nn as nn
 import torch.nn.functional as F
 import torch.optim as optim
 import torch_xla
-import torch_xla_py.data_parallel as dp
-import torch_xla_py.utils as xu
-import torch_xla_py.xla_model as xm
+import torch_xla.distributed.data_parallel as dp
+import torch_xla.utils.utils as xu
+import torch_xla.core.xla_model as xm
 import torchvision
 import torchvision.transforms as transforms
 import unittest
