+++ test/test_utils.py
@@ -1,8 +1,8 @@
 import os
 import sys
 import time
-import torch_xla_py.xla_model as xm
-import torch_xla_py.utils as xu
+import torch_xla.core.xla_model as xm
+import torch_xla.utils.utils as xu
 
 
 def _get_device_spec(device):
