+++ torch_xla/distributed/data_parallel.py
@@ -9,9 +9,9 @@ import threading
 import torch
 import torch.autograd
 import torch_xla
-import torch_xla_py.keyd_queue as kq
-import torch_xla_py.utils as xu
-import torch_xla_py.xla_model as xm
+import torch_xla.utils.keyd_queue as kq
+import torch_xla.utils.utils as xu
+import torch_xla.core.xla_model as xm
 import traceback
